pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "encontreras"
version = "0.1.0"
description = "CLI tool to extract business data from Google Maps with cascade enrichment and entity resolution"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.49.1",
    "pandas>=2.2.3",
    "typer[all]>=0.15.1",
    "rich>=13.9.4",
    "hupper>=1.12.1",
    "huey>=2.5.2",
    "google-genai>=0.3.0",
    "notion-client>=2.3.1",
    "python-dotenv>=1.0.1",
]

[project.scripts]
encontreras = "main:app"

[tool.setuptools.packages.find]
where = ["."]
